% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_de_imdb.R
\name{convert_DE_imdb}
\alias{convert_DE_imdb}
\title{Returns IMDB Details from German Movie Title(s)}
\usage{
convert_DE_imdb(movies_DE_EN = convert_movies(imdb_query =
  c("Love, Simon", "Vom Ende einer Geschichte")), query = NULL,
  return = "id")
}
\arguments{
\item{imdb_query}{}
}
\value{
an IMDB ID or URL
}
\description{
Returns IMDB Details from German Movie Title(s)
}
\examples{
convert_DE_imdb(query = "Love, Simon")  # "tt5164432"
convert_DE_imdb(query = "love, simon")  # error # case-senstive
convert_DE_imdb(query = "Love, Simon", return = "url")  # "https://www.imdb.com/title/tt5164432"

# should also work for non-English titles or multiple queries
convert_DE_imdb(query = c("Jurassic World: Das gefallene Königreich", "Vom Ende einer Geschichte"))  # "tt4827986" "tt4881806"
# convert_DE_imdb(query = c("Jurassic World: Das gefallene Königreich", "Vom Ende einer Geschichte"), return = "url")

\dontrun{
convert_DE_imdb()  # error
convert_DE_imdb(query = c("I am a movie", "Love, Simon"))  # warning; "tt5164432"
}
}
