% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_movies.R
\name{filter_movies}
\alias{filter_movies}
\title{Filter Movie Results}
\usage{
filter_movies(mov_info = get_mov_info_de(), details = NULL,
  quoted = FALSE)
}
\arguments{
\item{mov_info}{dataframe of movie information}

\item{details}{which details to retreive}

\item{quoted}{whether the details parameter should be quoted}
}
\value{
filtered movies
}
\description{
Filter Movie Results
}
\examples{
filter_movies(mov_info = get_mov_info_de())  # returns all details
filter_movies(mov_info = get_mov_info_en(), details = c("title", "runtime", "plot"))  # returns selected columns
filter_movies(details = expr(starts_with("poster")), quoted = TRUE)  ##TODO: check expr/quo/enquo
filter_movies(details = c("Who am I?", "Who are you?"))  # error
filter_movies(mov_info = get_mov_info_en(), details = c("title", "runtime", ":P", ":D"))  # warning; returns matchess
}
