% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_screenings.R
\name{filter_screenings}
\alias{filter_screenings}
\title{Filter Available Screenings}
\usage{
filter_screenings(screenings_details = get_screenings(),
  sel_dates = NULL, sel_release_types = NULL, sel_sites = NULL)
}
\arguments{
\item{screenings_details}{a data frame with screening details; output from `get_screenings()`}

\item{sel_dates}{selected date(s) in "YYYY-MM-DD" format as a vector of strings}

\item{sel_release_types}{selected release type(s) as a vector of strings}

\item{sel_sites}{selected site(s) as a vector of strings}
}
\value{
a tibble of movie titles (in German) and screening details
}
\description{
Filter Available Screenings
}
\examples{
filter_screenings(my_screenings_details)  # returns all screenings in my_screenings_details
filter_screenings(sel_dates = Sys.Date())  # returns all screenings found playing today
filter_screenings(sel_sites = c("Cineplex", "Schloßtheater"))  # returns all screenings found playing in Cineplex or Schloßtheater branches
filter_screenings(sel_release_types = "2D OmU")  # returns all screenings found playing in 2D OmU
filter_screenings(sel_sites = "just making stuff up now")  # "no results were found"
}
