% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mov_info_de.R
\name{get_mov_info_de}
\alias{get_mov_info_de}
\title{Retreive Movie Details from Cineplex}
\usage{
get_mov_info_de(city_html = get_city(city_url =
  "https://www.cineplex.de/filmreihe/originals/614/muenster/"))
}
\arguments{
\item{sel_imdb_ids}{}
}
\value{
details eg. runtime, poster URL..
}
\description{
Retreive Movie Details from Cineplex
}
\examples{
my_mov_info_DE <- get_mov_info_de()
my_mov_info_DE <- get_mov_info_de(city_html = get_city(city_url = "https://www.cineplex.de/filmreihe/originals/614/muenster/"))
}
