% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mov_info_en.R
\name{get_mov_info_en}
\alias{get_mov_info_en}
\title{Retreive Movie Details from IMDB}
\usage{
get_mov_info_en(sel_imdb_ids = c("tt5164432", "tt4881806"))
}
\arguments{
\item{sel_imdb_ids}{}
}
\value{
details eg. ratings, runtime, poster URL..
}
\description{
Retreive Movie Details from IMDB
}
\examples{
get_mov_info_en()
}
