% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_screenings.R
\name{get_screenings}
\alias{get_screenings}
\title{Retreive Screenings Available at a Cineplex City}
\usage{
get_screenings(city_html = get_city(city_url =
  "https://www.cineplex.de/filmreihe/originals/614/muenster/"))
}
\arguments{
\item{city_html}{}
}
\value{
input for filter_screenings()
}
\description{
Retreive Screenings Available at a Cineplex City
}
\examples{
my_screenings_details <- get_screenings(city_html = my_city)
}
