% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_movies.R
\name{convert_movies}
\alias{convert_movies}
\title{Returns Relevant Details from German Movie Title(s)}
\usage{
convert_movies(imdb_query = c("Love, Simon",
  "Vom Ende einer Geschichte"))
}
\arguments{
\item{imdb_query}{name of movie to search in IMDB as a vector of strings}
}
\value{
a tibble with English title(s) and IMDB details
}
\description{
Returns Relevant Details from German Movie Title(s)
}
\examples{
convert_movies(c("Love, Simon", "Vom Ende einer Geschichte"))
}
