% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_city.R
\name{get_city}
\alias{get_city}
\title{Retrieve Details from the Page of a Cineplex City}
\usage{
get_city(city_url)
}
\arguments{
\item{city_url}{}
}
\value{
input for get_screenings() (at the moment only testing Münster originals)
}
\description{
Retrieve Details from the Page of a Cineplex City
}
\examples{
my_city <- get_city(city_url = "https://www.cineplex.de/filmreihe/originals/614/muenster/")
}
